ensure cache directory exists before instantiating cache
Hilariously, Keyv only throws an error once you *access* the
cache, after instantiating it with a directory that doesn't exist.
If you instantiate it *and then* create the directory (e.g. before
accessing it), that *still* errors. So, we need to actually create
the directory before instantiating it.

We still reuse the same cache for subsequent calls though.